remove spoke health check from batch
